+++ src/test/compile-middleware_test.ts
@@ -18,7 +18,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as supertest from 'supertest';
 
-import {babelCompileCache, isPolyfill} from '../compile-middleware';
+import {babelCompileCache, getCompileCacheKey, isPolyfill} from '../compile-middleware';
 import {getApp} from '../start_server';
 
 chai.use(chaiAsPromised);
@@ -32,8 +32,11 @@ const userAgentsThatDontSupportES2015OrModules = [
   'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/52.0.2743.116 Safari/537.36 Edge/15.14986',
 ];
 
+const chrome61UA =
+    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/61.0.3163.39 Safari/537.36';
+
 const userAgentsThatSupportES2015AndModules = [
-  'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/61.0.3163.39 Safari/537.36',
+  chrome61UA,
   'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_6) AppleWebKit/603.3.8 (KHTML, like Gecko) Version/10.1.2 Safari/603.3.8',
 ];
 
@@ -50,10 +53,9 @@ suite('compile-middleware', () => {
         readTestFile('bower_components/test-component/test.html');
     const uncompiledJs =
         readTestFile('bower_components/test-component/test.js');
-    const optionsLeader =
-        JSON.stringify({transformES2015: true, transformModules: true});
+    const options = {transformES2015: true, transformModules: true};
 
-    beforeEach(() => {
+    setup(() => {
       app = getApp({
         root: root,
         compile: 'always',
@@ -64,68 +66,68 @@ suite('compile-middleware', () => {
     });
 
     test('caches html compilation results', async () => {
-      assert(!babelCompileCache.has(optionsLeader + uncompiledHtml));
-      const response =
-          await supertest(app).get('/components/test-component/test.html');
-      assert(babelCompileCache.has(optionsLeader + uncompiledHtml));
-      assert.equal(
-          response.text, babelCompileCache.get(optionsLeader + uncompiledHtml));
+      const requestPath = '/components/test-component/test.html';
+      const cacheKey = getCompileCacheKey(requestPath, uncompiledHtml, options);
+      assert.isFalse(babelCompileCache.has(cacheKey));
+      const response = await supertest(app).get(requestPath).expect(200);
+      assert.isTrue(babelCompileCache.has(cacheKey));
+      assert.equal(response.text, babelCompileCache.get(cacheKey));
       assert.equal(response.text.indexOf('class A {}'), -1, 'Did not compile');
     });
 
     test('returns cached html compilation results', async () => {
-      babelCompileCache.set(optionsLeader + uncompiledHtml, 'IM IN UR CACHE');
-      const response =
-          await supertest(app).get('/components/test-component/test.html');
+      const requestPath = '/components/test-component/test.html';
+      const cacheKey = getCompileCacheKey(requestPath, uncompiledHtml, options);
+      babelCompileCache.set(cacheKey, 'IM IN UR CACHE');
+      const response = await supertest(app).get(requestPath);
       assert.equal(response.text, 'IM IN UR CACHE');
     });
 
     test('caches javascript compilation results', async () => {
-      assert(!babelCompileCache.has(optionsLeader + uncompiledJs));
-      const response =
-          await supertest(app).get('/components/test-component/test.js');
-      assert(babelCompileCache.has(optionsLeader + uncompiledJs));
-      assert.equal(
-          response.text, babelCompileCache.get(optionsLeader + uncompiledJs));
+      const requestPath = '/components/test-component/test.js';
+      const cacheKey = getCompileCacheKey(requestPath, uncompiledJs, options);
+      assert.isFalse(babelCompileCache.has(cacheKey));
+      const response = await supertest(app).get(requestPath);
+      assert.isTrue(babelCompileCache.has(cacheKey));
+      assert.equal(response.text, babelCompileCache.get(cacheKey));
       assert.equal(response.text.indexOf('class A {}'), -1, 'Did not compile');
     });
 
     test('returns cached js compilation results', async () => {
-      babelCompileCache.set(optionsLeader + uncompiledJs, 'IM IN UR CACHE');
-      const response =
-          await supertest(app).get('/components/test-component/test.js');
+      const requestPath = '/components/test-component/test.js';
+      const cacheKey = getCompileCacheKey(requestPath, uncompiledJs, options);
+      babelCompileCache.set(cacheKey, 'IM IN UR CACHE');
+      const response = await supertest(app).get(requestPath);
       assert.equal(response.text, 'IM IN UR CACHE');
     });
 
     test('honors the cache max evicting least recently used', async () => {
-      await supertest(app).get('/components/test-component/test.html');
-      assert(babelCompileCache.has(optionsLeader + uncompiledHtml));
+      const requestPath = '/components/test-component/test.html';
+      const cacheKey = getCompileCacheKey(requestPath, uncompiledHtml, options);
+      await supertest(app).get(requestPath);
+      assert.isTrue(babelCompileCache.has(cacheKey));
       const originalMax = babelCompileCache['max'];
       babelCompileCache['max'] = babelCompileCache.length;
       try {
+        // request a file not in the cache
         await supertest(app).get('/components/test-component/test.js');
-        assert(
-            !babelCompileCache.has(optionsLeader + uncompiledHtml),
-            'cached html evicted');
+        assert.isFalse(babelCompileCache.has(cacheKey), 'cached html evicted');
       } finally {
         babelCompileCache['max'] = originalMax;
       }
     });
 
     test('script tags with invalid javascript are unchanged', async () => {
-      const uncompiled =
-          fs
-              .readFileSync(path.join(
-                  root, 'bower_components/compile-test/script-tags.html'))
-              .toString();
-      assert(
-          !babelCompileCache.has(optionsLeader + uncompiled),
-          'Unexpected entry in cache');
-      const response =
-          await supertest(app).get('/components/compile-test/script-tags.html');
-      assert(
-          babelCompileCache.has(optionsLeader + uncompiled),
-          'Missing cache entry');
+      const filePath =
+          path.join(root, 'bower_components/compile-test/script-tags.html');
+      const requestPath = '/components/compile-test/script-tags.html';
+      const uncompiled = fs.readFileSync(filePath).toString();
+      const cacheKey = getCompileCacheKey(requestPath, uncompiled, options);
+      assert.isFalse(
+          babelCompileCache.has(cacheKey), 'Unexpected entry in cache');
+
+      const response = await supertest(app).get(requestPath);
+      assert(babelCompileCache.has(cacheKey), 'Missing cache entry');
       assert.include(response.text, `<script>\nthis is not valid\n</script>`);
       // The valid script tag should still be compiled.
       assert.notInclude(response.text, `<script>\nclass A {}\n</script>`);
@@ -133,7 +135,7 @@ suite('compile-middleware', () => {
 
     suite('with compile option set to \'auto\'', () => {
 
-      beforeEach(() => {
+      setup(() => {
         app = getApp({
           root: root,
           compile: 'auto',
@@ -144,13 +146,16 @@ suite('compile-middleware', () => {
       });
 
       test('detect user-agents that do not need compilation', async () => {
-        assert.isFalse(babelCompileCache.has(`Unexpected .js file in cache`));
+        const requestPath = '/components/test-component/test.js';
+        const cacheKey = getCompileCacheKey(requestPath, uncompiledJs, options);
+        assert.isFalse(
+            babelCompileCache.has(cacheKey), `Unexpected .js file in cache`);
         for (const userAgent of userAgentsThatSupportES2015AndModules) {
           const response = await supertest(app)
                                .get('/components/test-component/test.js')
                                .set('User-Agent', userAgent);
           assert.isFalse(
-              babelCompileCache.has(optionsLeader + uncompiledJs),
+              babelCompileCache.has(cacheKey),
               `Unexpected .js file in cache User-Agent ${userAgent}`);
           assert.include(
               response.text,
@@ -161,13 +166,16 @@ suite('compile-middleware', () => {
       });
 
       test('detect user-agents that need compilation', async () => {
-        assert.isFalse(babelCompileCache.has(`Unexpected .js file in cache`));
+        const requestPath = '/components/test-component/test.js';
+        const cacheKey = getCompileCacheKey(requestPath, uncompiledJs, options);
+        assert.isFalse(
+            babelCompileCache.has(cacheKey), `Unexpected .js file in cache`);
         for (const userAgent of userAgentsThatDontSupportES2015OrModules) {
           const response = await supertest(app)
                                .get('/components/test-component/test.js')
                                .set('User-Agent', userAgent);
           assert.isTrue(
-              babelCompileCache.has(optionsLeader + uncompiledJs),
+              babelCompileCache.has(cacheKey),
               `Expected .js file in cache User-Agent ${userAgent}`);
           assert.notInclude(
               response.text,
@@ -198,7 +206,7 @@ suite('compile-middleware', () => {
     const userAgent =
         'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3163.39 Safari/537.36';
 
-    beforeEach(() => {
+    setup(() => {
       app = getApp({
         root: root,
         compile: 'auto',
@@ -242,5 +250,55 @@ suite('compile-middleware', () => {
       assert.equal(response.status, 200);
       assert.include(response.text, 'requirejs');
     });
+
   });
+
+  suite('module transformations', () => {
+    const userAgent = chrome61UA;
+
+    setup(() => {
+      app = getApp({
+        root: root,
+        compile: 'never',
+        moduleResolution: 'node',
+        componentDir: path.join(root, 'node_modules'),
+      });
+      babelCompileCache.reset();
+    });
+
+    async function assertGolden(
+        filename: string, componentRequest: boolean, index = false) {
+      let filePath = path.join('node_modules/golden', filename);
+      if (index) {
+        filePath = path.join(filePath, 'index.html');
+      }
+      const golden = readTestFile(filePath);
+      const response =
+          await supertest(app)
+              .get(
+                  path.join((componentRequest ? '/components/' : ''), filename))
+              .set('User-Agent', userAgent);
+      assert.equal(response.text.trim(), golden.trim());
+    }
+
+    test('rewrites bare specifiers in component requests', async () => {
+      await assertGolden('@polymer/test-component/test-module.js', true);
+    });
+
+    test('rewrites bare specifiers in HTML in component requests', async () => {
+      await assertGolden('@polymer/test-component/index.html', true);
+    });
+
+    test(
+        'rewrites bare specifiers in index.html in component requests',
+        async () => {
+          await assertGolden('@polymer/test-component/', true, true);
+        });
+
+    test('rewrites bare specifiers in app requests', async () => {
+      await assertGolden('/app.js', false);
+    });
+
+  });
+
 });